refactor: add Auth scope to auth related Jwt logics
