+++ app/commands/manager/set_user_manager_tokens.rb
@@ -14,7 +14,7 @@ module Manager
       with_exception_handle do
         response = Request::ApiRequest.get(google_api_url)
         @user_manager = UserManager.create_user_for_google(response)
-        access_token, refresh_token = Jwt::Issuer.call(@user_manager)
+        access_token, refresh_token = Jwt::Auth::Issuer.call(@user_manager)
         @user_manager.save
 
         { access_token:, refresh_token: }
