+++ app/lib/jwt/auth/expiry.rb
@@ -1,0 +1,11 @@
+module Jwt
+  module Auth
+    module Expiry
+      module_function
+
+      def expiry
+        2.hours
+      end
+    end
+  end
+end
