+++ app/lib/jwt/auth/secret.rb
@@ -1,0 +1,11 @@
+module Jwt
+  module Auth
+    module Secret
+      module_function
+
+      def secret
+        Rails.application.secrets.secret_key_base
+      end
+    end
+  end
+end
