--- app/lib/jwt/expiry.rb
@@ -1,9 +1,0 @@
-module Jwt
-  module Expiry
-    module_function
-
-    def expiry
-      2.hours
-    end
-  end
-end
