--- app/lib/jwt/secret.rb
@@ -1,9 +1,0 @@
-module Jwt
-  module Secret
-    module_function
-
-    def secret
-      Rails.application.secrets.secret_key_base
-    end
-  end
-end
